app/requirements.txt
environ
Django
django-environ
djangorestframework
django-filter
psycopg2
django-imagekit